requirements.txt
# Core RAG dependencies
groq>=0.9.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0

# Evaluation metrics
rouge-score>=0.1.2
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0

# LangChain (optional, for extended pipeline experiments)
langchain>=0.2.0
langchain-groq>=0.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Jupyter demo
jupyterlab>=4.2.0
ipywidgets>=8.1.0
matplotlib>=3.9.0
seaborn>=0.13.0